feat: implement reusable publish modal tests and mock confirmation modal
SonarQube comment - Duplicated Lines
